Update hyper package to version 0.8.0 and enhance Python compatibility
This commit updates the hyper package to version 0.8.0 with necessary changes in imports and code revisions. The package is now compatible with Python 3.8 to 3.12, as prior versions are no longer supported. A few string patterns in `rfc3986/misc.py` and the README section have been amended accordingly. Hostname checks are no longer needed as ssl in >=3.8 does them as part of the handshake.